0.3.1.123 - Removido trycatch em ProductPackageService, as exceptions hoje são tratadas em ApiListener. - Movido algumas verificações de ProductPackageService para ProductPackageControl. - Ao excluir embalagens de produto será verificado anteriormente a existência de dependências. - Substituído alguns ApiResult por ApiResultObject em ProductPackageService. - Movido scripts de teste dos webservices para a pasta 'product-package-service'. - Agora as exceções geradas ao gerenciar ProductPackage terá um status diferente para cada problema. - Atualizado 'product_pckages' para que 'name' seja UNIQUE KEY. 	- Atualizado para verificações de unique key au adicionar e atualizar.

+++ src/tercom/dao/ProductPackageDAO.php
@@ -5,25 +5,79 @@ namespace tercom\dao;
 use tercom\entities\ProductPackage;
 use tercom\entities\lists\ProductPackages;
 use dProject\MySQL\Result;
+use tercom\exceptions\ProductPackageException;
+use tercom\exceptions\ProductPriceException;
 
+/**
+ * DAO para Embalagem de Produto
+ *
+ * Classe responsável pela comunicação completa do sistema para com o banco de dados.
+ * Sua responsabilidade é gerenciar os dados referentes as embalagens de produto, incluindo todas operações.
+ * Estas operações consiste em: adicionar, atualizar, excluir e selecionar dados das embalagens de produto.
+ *
+ * Cada embalagem deve possuir um nome sendo obrigatório e único no sisstema.
+ *
+ * @see GenericDAO
+ * @see ProductPackage
+ * @see ProductPackages
+ *
+ * @author Andrew
+ */
 class ProductPackageDAO extends GenericDAO
 {
+	/**
+	 * @var array nome das colunas da tabela de embalagens de produto.
+	 */
+	public const ALL_COLUMNS = ['id', 'name'];
+
+	/**
+	 * Procedimento interno para validação dos dados de uma embalagem de produto ao inserir e/ou atualizar.
+	 * Embalagens de produto devem possuir um nome e este deve ser único no sistema.
+	 * @param ProductPackage $productPackage objeto do tipo emblagem de produto à ser validado.
+	 * @param bool $validateId true para validar o código de identificação único ou false caso contrário.
+	 * @throws ProductPackageException caso algum dos dados da embalagem de produto não estejam de acordo.
+	 */
+	private function validate(ProductPackage $productPackage, bool $validateID)
+	{
+		// PRIMARY KEY
+		if ($validateID) {
+			if ($productPackage->getID() === 0)
+				throw ProductPackageException::newNotIdentified();
+		} else {
+			if ($productPackage->getID() !== 0)
+				throw ProductPackageException::newIdentified();
+		}
+
+		// UNIQUE KEY
+		if ($this->existName($productPackage->getName(), $productPackage->getId())) throw ProductPackageException::newNameUnavaiable();
+	}
+
+	/**
+	 * Insere uma nova embalagem de produto no banco de dados e atualiza o mesmo com o identificador gerado.
+	 * @param ProductPackage $productPackage objeto do tipo embalagem de produto à adicionar.
+	 * @return bool true se conseguir adicionar ou false caso contrário.
+	 */
 	public function insert(ProductPackage $productPackage): bool
 	{
+		$this->validate($productPackage, false);
+
 		$sql = "INSERT INTO product_packages (name)
 				VALUES (?)";
 
 		$query = $this->createQuery($sql);
 		$query->setString(1, $productPackage->getName());
 
-		$result = $query->execute();
-
-		if ($result->isSuccessful())
+		if (($result = $query->execute())->isSuccessful())
 			$productPackage->setID($result->getInsertID());
 
 		return $result->isSuccessful();
 	}
 
+	/**
+	 * Atualiza os dados de uma embalagem de produto já existente no banco de dados.
+	 * @param ProductPackage $productPackage objeto do tipo embalagem de produto à atualizar.
+	 * @return bool true se for atualizado ou false caso contrário.
+	 */
 	public function update(ProductPackage $productPackage): bool
 	{
 		$sql = "UPDATE product_packages
@@ -34,24 +88,34 @@ class ProductPackageDAO extends GenericDAO
 		$query->setString(1, $productPackage->getName());
 		$query->setInteger(2, $productPackage->getID());
 
-		$result = $query->execute();
-
-		return $result->isSuccessful();
+		return ($query->execute())->isSuccessful();
 	}
 
+	/**
+	 * Exclui uma embalagem de produto do banco de dados e considera utilizações.
+	 * Caso esteja sendo referenciada em outra tabela não será possível excluir.
+	 * @param ProductPackage $productPackage objeto do tipo contato de fornecedor.
+	 * @return bool true se excluído ou false caso contrário.
+	 */
 	public function dalete(ProductPackage $productPackage): bool
 	{
+		if ($this->existOnProductPrice($productPackage->getId()))
+			throw ProductPriceException::newHasUses();
+
 		$sql = "DELETE FROM product_packages
 				WHERE id = ?";
 
 		$query = $this->createQuery($sql);
 		$query->setInteger(1, $productPackage->getID());
 
-		$result = $query->execute();
-
-		return $result->getAffectedRows();
+		return ($query->execute())->getAffectedRows() === 1;
 	}
 
+	/**
+	 * Selecione os dados de uma embalagem de produto através do seu código de identificação único.
+	 * @param int $idProductPackage código de identificação único da embalagem de produto.
+	 * @return ProductPackage|NULL embalagem de produto com os dados carregados ou NULL se não encontrado.
+	 */
 	public function select(int $idProductPackage): ?ProductPackage
 	{
 		$sql = "SELECT id, name
@@ -66,6 +130,10 @@ class ProductPackageDAO extends GenericDAO
 		return $this->parseProductPackage($result);
 	}
 
+	/**
+	 * Seleciona os dados de todas as embalagens de produto registradas no banco de dados sem ordenação.
+	 * @return ProductPackages aquisição da lista de emblagens de produto atualmente registrados.
+	 */
 	public function selectAll(): ProductPackages
 	{
 		$sql = "SELECT id, name
@@ -77,7 +145,12 @@ class ProductPackageDAO extends GenericDAO
 		return $this->parseProductPackages($result);
 	}
 
-	public function searchByName(string $name): ProductPackages
+	/**
+	 * Seleciona os dados das embalagens de produto no banco de dados filtrados pelo nome.
+	 * @param string $name nome da emblagem parcial ou completo para filtro.
+	 * @return ProductPackages aquisição da lista de embalagens de produto conforme filtro.
+	 */
+	public function selectLikeName(string $name): ProductPackages
 	{
 		$sql = "SELECT id, name
 				FROM product_packages
@@ -91,24 +164,33 @@ class ProductPackageDAO extends GenericDAO
 		return $this->parseProductPackages($result);
 	}
 
-	public function existID(int $idProductPackage): bool
+	/**
+	 * Verifica se um determinado código de identificação de emblagem de produto existe.
+	 * @param int $idProductPackage código de identificação único da embalagem de produto.
+	 * @return bool true se existir ou false caso contrário.
+	 */
+	public function exist(int $idProductPackage): bool
 	{
-		$sql = "SELECT COUNT(*) AS qtd
+		$sql = "SELECT COUNT(*) qty
 				FROM product_packages
 				WHERE id = ?";
 
 		$query = $this->createQuery($sql);
 		$query->setInteger(1, $idProductPackage);
 
-		$result = $query->execute();
-		$productPackages = $result->next();
-
-		return intval($productPackages['qtd']) === 1;
+		return $this->parseQueryExist($query);
 	}
 
+	/**
+	 * Verifica se um nome de embalagem de produto está disponível par auso.
+	 * @param string $name nome da emblagem de produto à verificar.
+	 * @param int $idProductPackage código de identificação único da emblagem de produto
+	 * ou zero caso seja um novo fornecedor.
+	 * @return bool true se existir ou false caso contrário.
+	 */
 	public function existName(string $name, int $idProductPackage): bool
 	{
-		$sql = "SELECT COUNT(*) AS qtd
+		$sql = "SELECT COUNT(*) qty
 				FROM product_packages
 				WHERE name = ? AND id <> ?";
 
@@ -116,12 +198,31 @@ class ProductPackageDAO extends GenericDAO
 		$query->setString(1, $name);
 		$query->setInteger(2, $idProductPackage);
 
-		$result = $query->execute();
-		$productPackages = $result->next();
-
-		return intval($productPackages['qtd']) === 1;
+		return $this->parseQueryExist($query);
 	}
 
+	/**
+	 * Verifica se uma embalagem de produto está sendo referenciada em algum preço de produto.
+	 * @param int $idProductPackage código de identificação único da emblagem de produto.
+	 * @return bool true se existir ou false caso contrário.
+	 */
+	public function existOnProductPrice(int $idProductPackage): bool
+	{
+		$sql = "SELECT COUNT(*) qty
+				FROM product_prices
+				WHERE idProductPackage = ?";
+
+		$query = $this->createQuery($sql);
+		$query->setInteger(1, $idProductPackage);
+
+		return $this->parseQueryExist($query);
+	}
+
+	/**
+	 * Procedimento interno para analisar o resultado de uma consulta e criar um objeto de embalagem de produto.
+	 * @param Result $result referência do resultado da consulta obtido.
+	 * @return ProductPackage|NULL objeto do tipo embalagem de produto com dados carregados ou NULL se não houver resultado.
+	 */
 	private function parseProductPackage(Result $result): ?ProductPackage
 	{
 		if (!$result->hasNext())
@@ -133,6 +234,11 @@ class ProductPackageDAO extends GenericDAO
 		return $productPackage;
 	}
 
+	/**
+	 * Procedimento inerno para analisar o resultado de uma consulta e criar os objetos de embalagem de produto.
+	 * @param Result $result referência do resultado da consulta obtido.
+	 * @return ProductPackages aquisição da lista de embalagens de produto a partir da consulta.
+	 */
 	private function parseProductPackages(Result $result): ProductPackages
 	{
 		$productPackages = new ProductPackages();
@@ -147,6 +253,11 @@ class ProductPackageDAO extends GenericDAO
 		return $productPackages;
 	}
 
+	/**
+	 * Procedimento interno para criar um objeto do tipo embalagem de produto e carregar os dados de um registro.
+	 * @param array $entry vetor contendo os dados do registro obtido de uma consulta.
+	 * @return ProductPackage aquisição de um objeto do tipo embalagem de produto com dados carregados.
+	 */
 	private function newProductPackage(array $array): ProductPackage
 	{
 		$productPackage = new ProductPackage();
